update: extract express download management
