+++ backend/src/controllers/object.ts
@@ -8,6 +8,7 @@ import {
 } from '../useCases/index.js'
 import { logger } from '../drivers/logger.js'
 import { asyncSafeHandler } from '../utils/express.js'
+import { handleDownloadResponseHeaders } from '../services/download/express.js'
 
 const objectController = Router()
 
@@ -219,16 +220,7 @@ objectController.get(
         return
       }
 
-      const safeName = encodeURIComponent(metadata.name || 'download')
-
-      if (metadata.type === 'file') {
-        res.set('Content-Type', metadata.mimeType || 'application/octet-stream')
-        res.set('Content-Disposition', `attachment; filename="${safeName}"`)
-        res.set('Content-Length', metadata.totalSize.toString())
-      } else {
-        res.set('Content-Type', 'application/zip')
-        res.set('Content-Disposition', `attachment; filename="${safeName}.zip"`)
-      }
+      handleDownloadResponseHeaders(res, metadata)
 
       pipeline(await startDownload(), res, (err) => {
         if (err) {
@@ -347,21 +339,7 @@ objectController.get(
         return
       }
 
-      const safeName = encodeURIComponent(metadata.name || 'download')
-      if (metadata.type === 'file') {
-        res.set('Content-Type', metadata.mimeType || 'application/octet-stream')
-        res.set('Content-Disposition', `filename="${safeName}"`)
-        res.set('Content-Length', metadata.totalSize.toString())
-        const compressedButNoEncrypted =
-          metadata.uploadOptions?.compression &&
-          !metadata.uploadOptions?.encryption
-        if (compressedButNoEncrypted) {
-          res.set('Content-Encoding', 'deflate')
-        }
-      } else {
-        res.set('Content-Type', 'application/zip')
-        res.set('Content-Disposition', `filename="${safeName}.zip"`)
-      }
+      handleDownloadResponseHeaders(res, metadata)
 
       pipeline(await startDownload(), res, (err) => {
         if (err) {
